refactor: extract generic AppLayout shell and shared WalletButton
Replace duplicated header/nav/footer structure across layouts with a
single AppLayout component that accepts logo, navItems, isActive,
headerRight, and mobileExtra props. Extract wallet connect logic into
a standalone WalletButton component (no visibility classes) so it can
be reused in both desktop and mobile slots across all layouts.
NavbarWallet becomes a thin hidden-md:flex wrapper around WalletButton.
