+++ hooks/useActiveNavigation.ts
@@ -4,7 +4,7 @@ export function useActiveNavigation() {
   const router = useRouter();
   
   const isActive = (path: string) => {
-    if (path === '/dashboard' || path === '/deribit-agent') {
+    if (path === '/dashboard' || path === '/deribit-agent' || path === '/routes') {
       return router.pathname === path;
     }
     return router.pathname.startsWith(path);
